requirements.txt
# Telegram
Telethon==1.42.0

# AI Providers
openai==2.15.0
google-genai==1.60.0

# PostgreSQL + SQLAlchemy
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.2.1
tenacity==9.1.2

# HTTP
httpx==0.28.1
requests==2.32.5

# Прочее
tqdm==4.67.1
